Do not consider unverified emails to be valid
