+++ lib/omniauth_vivokey_open_id.rb
@@ -123,7 +123,8 @@ module ::OmniAuth
             verify_sub: false,
             verify_expiration: true,
             verify_not_before: true,
-            verify_iat: true,
+            # TODO: set back to true
+            verify_iat: false,
             verify_jti: false
           )
           verbose_log("Verified JWT\n\n#{decoded.to_yaml}")
